+++ src/main/java/com/oheat/user/dto/CartSaveRequest.java
@@ -1,8 +1,8 @@
-package com.oheat.order.dto;
+package com.oheat.user.dto;
 
 import com.oheat.food.entity.MenuJpaEntity;
 import com.oheat.food.entity.ShopJpaEntity;
-import com.oheat.order.entity.CartJpaEntity;
+import com.oheat.user.entity.CartJpaEntity;
 import com.oheat.user.entity.UserJpaEntity;
 import java.util.List;
 import lombok.Builder;
