+++ src/main/java/com/oheat/user/exception/CartNotExistsException.java
@@ -1,4 +1,4 @@
-package com.oheat.order.exception;
+package com.oheat.user.exception;
 
 public class CartNotExistsException extends RuntimeException {
