+++ src/main/java/com/oheat/user/exception/OtherShopMenuAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.oheat.order.exception;
+package com.oheat.user.exception;
 
 public class OtherShopMenuAlreadyExistsException extends RuntimeException {
