+++ src/main/java/com/oheat/user/repository/CartRepository.java
@@ -1,7 +1,7 @@
-package com.oheat.order.repository;
+package com.oheat.user.repository;
 
 import com.oheat.food.entity.MenuJpaEntity;
-import com.oheat.order.entity.CartJpaEntity;
+import com.oheat.user.entity.CartJpaEntity;
 import com.oheat.user.entity.UserJpaEntity;
 import java.util.List;
 import java.util.Optional;
