+++ src/main/java/com/oheat/user/service/CartService.java
@@ -1,4 +1,4 @@
-package com.oheat.order.service;
+package com.oheat.user.service;
 
 import com.oheat.food.entity.MenuJpaEntity;
 import com.oheat.food.entity.OptionJpaEntity;
@@ -11,17 +11,17 @@ import com.oheat.food.repository.MenuRepository;
 import com.oheat.food.repository.OptionGroupRepository;
 import com.oheat.food.repository.OptionRepository;
 import com.oheat.food.repository.ShopRepository;
-import com.oheat.order.dto.CartSaveRequest;
-import com.oheat.order.dto.CartSaveRequest.CartOptionGroupSaveRequest;
-import com.oheat.order.entity.CartJpaEntity;
-import com.oheat.order.entity.CartOptionGroup;
-import com.oheat.order.entity.CartOptionGroupOption;
-import com.oheat.order.exception.CannotCreateCartException;
-import com.oheat.order.exception.CartNotExistsException;
-import com.oheat.order.exception.OtherShopMenuAlreadyExistsException;
-import com.oheat.order.repository.CartRepository;
+import com.oheat.user.dto.CartSaveRequest;
+import com.oheat.user.dto.CartSaveRequest.CartOptionGroupSaveRequest;
+import com.oheat.user.entity.CartJpaEntity;
+import com.oheat.user.entity.CartOptionGroup;
+import com.oheat.user.entity.CartOptionGroupOption;
 import com.oheat.user.entity.UserJpaEntity;
+import com.oheat.user.exception.CannotCreateCartException;
+import com.oheat.user.exception.CartNotExistsException;
+import com.oheat.user.exception.OtherShopMenuAlreadyExistsException;
 import com.oheat.user.exception.UserNotExistsException;
+import com.oheat.user.repository.CartRepository;
 import com.oheat.user.repository.UserRepository;
 import java.util.Comparator;
 import java.util.List;
