+++ src/test/java/com/oheat/user/MemoryCartRepository.java
@@ -1,9 +1,9 @@
-package com.oheat.order;
+package com.oheat.user;
 
 import com.oheat.food.entity.MenuJpaEntity;
-import com.oheat.order.entity.CartJpaEntity;
-import com.oheat.order.repository.CartRepository;
+import com.oheat.user.entity.CartJpaEntity;
 import com.oheat.user.entity.UserJpaEntity;
+import com.oheat.user.repository.CartRepository;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
